pip-requirements.txt
basemap
matplotlib
numpy
scipy
cpc.stats
cpc.geogrids
cpc.geoplot
